move entity const to entity file
