+++ src/packages/user/user/entity-bulk-actions/manifests.ts
@@ -4,7 +4,7 @@ import {
 	UMB_UNLOCK_USER_REPOSITORY_ALIAS,
 	UMB_USER_REPOSITORY_ALIAS,
 } from '../repository/manifests.js';
-import { UMB_USER_ENTITY_TYPE } from '../types.js';
+import { UMB_USER_ENTITY_TYPE } from '../entity.js';
 import { UMB_USER_COLLECTION_ALIAS } from '../collection/manifests.js';
 import { UmbEnableUserEntityBulkAction } from './enable/enable.action.js';
 import { UmbSetGroupUserEntityBulkAction } from './set-group/set-group.action.js';
