+++ src/packages/user/user/entity.ts
@@ -1,0 +1,1 @@
+export const UMB_USER_ENTITY_TYPE = 'user';
