+++ src/packages/user/user/invite/entity-action/manifests.ts
@@ -1,5 +1,5 @@
 import { UMB_INVITE_USER_REPOSITORY_ALIAS } from '../repository/manifests.js';
-import { UMB_USER_ENTITY_TYPE } from '../../index.js';
+import { UMB_USER_ENTITY_TYPE } from '../../entity.js';
 import { UmbResendInviteToUserEntityAction } from './resend-invite/resend-invite.action.js';
 import { manifest as conditionManifest } from './resend-invite/resend-invite.action.condition.js';
 import { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
