+++ src/packages/user/user/workspace/user-workspace.context.ts
@@ -1,4 +1,5 @@
-import { UMB_USER_ENTITY_TYPE, type UmbUserDetailModel } from '../types.js';
+import { type UmbUserDetailModel } from '../types.js';
+import { UMB_USER_ENTITY_TYPE } from '../entity.js';
 import { UmbUserDetailRepository } from '../repository/index.js';
 import {
 	UmbSaveableWorkspaceContextInterface,
